♻️ Make it so the bot can reply with multiple messages
Until now, the bot could only reply to a message with
a single message. It was basically request/response.

Now the bot can send multiple replies to a request.
If for example, your bot is asked to check the status
of 100 IoT devices, it can now reply with the status
for each device as the results come in instead of having
to wait until all devices have responded.